requirements.txt
setuptools~=50.3.0
psutil~=5.7.2
SQLAlchemy~=1.3.19
pandas~=1.1.2
numpy~=1.19.2
scipy~=1.5.2
PyYAML~=5.3.1
PyMySQL~=0.10.1
SQLAlchemy-Utils~=0.36.8